Form creation

+++ djangoadmin/app1/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class User(models.Model):
   name = models.CharField('name', max_length=30)
-  phone = models.IntegerField('phone')
+  phone = models.BigIntegerField('phone')
   email = models.CharField('email', max_length=30)
 
   def __srt__(self):
